Move testes AdicionarEtapaView para arquivo próprio #63
